backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend Component"
authors = ["miketvo <miketvo@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pandas = "*"
pillow = "9.5.0"
scikit-learn = "1.2.2"
tensorflow = "2.12.0"
tensorflow-addons = "0.20.0"
tensorflow-io-gcs-filesystem = "0.27.0"
keras = "2.12.0"
keras-tuner = "1.3.5"
numpy = "1.23.5"
joblib = "*"
fastapi = "*"
uvicorn = "^0.24.0.post1"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"


[tool.poetry.scripts]
start = "backend.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
